requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
customtkinter==5.1.3
darkdetect==0.8.0
dataclasses-json==0.5.7
frozenlist==1.3.3
geographiclib==2.0
geopy==2.3.0
greenlet==2.0.2
idna==3.4
langchain==0.0.171
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==0.4.3
numexpr==2.8.4
numpy==1.24.3
openapi-schema-pydantic==1.2.4
opencv-contrib-python==4.7.0.72
packaging==23.0
pydantic==1.10.2
PyYAML==6.0
requests==2.30.0
SQLAlchemy==2.0.13
tenacity==8.2.2
typing-inspect==0.7.1
typing_extensions==4.5.0
urllib3==2.0.2
yarl==1.9.2
Pillow==9.5.0
